requirements.txt
django
djangorestframework == 3.9.2